renamed services files and service spec.ts files. changes documented in GoogleDoc
