+++ CaliberTrainerReport/src/app/Components/assess-scores-by-category/assessment-scores-according-to-category.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule } from '@angular/forms';
 import { of } from 'rxjs';
-import { AssessmentByCategoryService } from 'src/app/services/AssessmentByCategory.service';
+import { AssessmentByCategoryService } from 'src/app/services/assess-by-category.service';
 import { DisplayGraphService } from 'src/app/services/display-graph.service';
 import { AssessmentComponent } from '../assessment/assessment.component';
