+++ CaliberTrainerReport/src/app/Components/assess-scores-by-category/assessment-scores-according-to-category.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, HostListener, OnDestroy } from '@angular/core';
 import { faChartBar, faTable } from '@fortawesome/free-solid-svg-icons';
 import { Chart } from 'node_modules/chart.js';
-import { AssessmentByCategoryService } from 'src/app/services/AssessmentByCategory.service';
+import { AssessmentByCategoryService } from 'src/app/services/assess-by-category.service';
 import { AssessmentComponent } from 'src/app/Components/assessment/assessment.component';
 import { Subscription } from 'rxjs';
 import { DisplayGraphService } from 'src/app/services/display-graph.service';
