+++ CaliberTrainerReport/src/app/Components/qc-week-scores/qc-overall-week-technical-scores.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, HostListener, OnDestroy } from '@angular/core';
 import { faChartBar } from '@fortawesome/free-solid-svg-icons';
 import { Chart } from 'node_modules/chart.js';
-import { TechnicalStatusByWeekService } from 'src/app/services/TechnicalStatusByWeek.service';
+import { TechnicalStatusByWeekService } from 'src/app/services/tech-status-by-week.service';
 import { QCComponent } from 'src/app/Components/qc/qc.component';
 import { Subscription } from 'rxjs';
 import { DisplayGraphService } from 'src/app/services/display-graph.service';
