+++ CaliberTrainerReport/src/app/app.module.ts
@@ -4,7 +4,7 @@ import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { FormsModule } from '@angular/forms';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { TechnicalStatusPerBatchService } from './services/TechnicalStatusPerBatch.service';
+import { TechnicalStatusPerBatchService } from './services/tech-status-per-batch.service';
 import { MainNavBarComponent } from './Components/main-nav-bar/main-nav-bar.component';
 import { QcOverallBatchTechnicalScoresComponent } from './Components/qc-batch-scores/qc-overall-batch-technical-scores.component';
 import { QcOverallWeekTechnicalScoresComponent } from './Components/qc-week-scores/qc-overall-week-technical-scores.component';
@@ -13,9 +13,9 @@ import { AssessmentScoresForCategoryComponent } from './Components/filter-scores
 import { QCComponent } from './Components/qc/qc.component';
 import { AssessmentComponent } from './Components/assessment/assessment.component';
 import { HomeComponent } from './Components/home/home.component';
-import { BatchTechnicalStatusBySkillCategoryService } from './services/BatchTechnicalStatusBySkillCategory.service';
-import { AssessmentByBatchService } from './services/AssessmentByBatch.service';
-import { AssessmentByCategoryService } from './services/AssessmentByCategory.service';
+import { BatchTechnicalStatusBySkillCategoryService } from './services/batch-status-by-skill-cat.service';
+import { AssessmentByBatchService } from './services/assess-by-batch.service';
+import { AssessmentByCategoryService } from './services/assess-by-category.service';
 import { FileUploadService } from './services/file-upload.service';
 import { UrlService } from './services/url.service';
 import { HttpClientModule } from '@angular/common/http';
