+++ CaliberTrainerReport/src/app/services/batch-status-by-skill-cat.service.ts
