+++ CaliberTrainerReport/src/app/services/scores-by-category-all-batches.service.ts
